+++ packages/navi.js/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @awell-health/navi-dot-js
 
+## 0.0.5-beta.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @awell-health/navi-core@0.0.5-beta.3
+
 ## 0.0.5-beta.2
 
 ### Patch Changes